Added `Avada` theme and `Fusion Builder` plugin for the LMS. EW-361
Since we're using some custom Fusion Builder styling the Avada and Fusion Builder support libraries were brought in from the Wordpress marketing frontend. 

Todo: Need to consider whether or not this works well going forward. At the moment it is only applied to the LMS non-logged in pages (home / login) and logged in page (dashboard). Have not tested this with the Studio application.