Initial setup for `badges` as new courseware tab. EW-405
We need to create a new Django app that is integrated into the platform since it uses base class CourseTab that is currently not setup to be used outside of the platform code. The code currently just makes the plugin setup but does not define calls to integrate badging-frontend code at the moment.

This code update defines #4 from this page by adding new Django app to the platform codebase.
https://openedx.atlassian.net/wiki/spaces/AC/pages/30965856/How+to+add+a+new+feature+to+LMS+or+Studio