feat: Fixed issue with `CourseBadgesTab` is_enabled check. EW-525
When serializing the `tabs` for field for `lms.djangoapps.courseware.tabs.CourseInfoTab` there was an error retrieving the `issue_badges` property for the `course` instance for `CourseBadgesTab` since the instance was a `CourseOverview` class. That field is found from the `CourseDescriptorWithMixins` class only for that course advanced field setting. This commit fixes that issue.