fix(bigcommerce): Resolve issue when BigCommerce Product doesn't include custom field `Course ID`
Need to make sure that the `custom_fields` returned from the BigCommerce product include the `Course ID` field. This does a check to see if the BigCommerce client `ProductCustomFields` exists before checking the value since this produces and `str` result for `field` value whenever the Custom Field is not setup.