fix(configuration): Only import `private.py` in production conditionally.
When running devstack environment this `./envs/production.py` import for the `private.py` needs to not be included in this file because it is imported a second time at the end of `./envs/devstack.py`. Having it imported twice it will only import the first `./envs/private.py` file. This causes issues when you think that you're overriding a FEATURE flag and `./envs/devstack.py` updates that flag after the `./envs/private.py` has already been imported.