feat: Extend openedx `announcements` feature to distinguish banner from sidebar notifications.
The openedx feature for announcements already exists. We extended this allow for `banner` and `sidebar` announcements from the LMS course dashboard page.

We needed a way to notify the learner about upcoming maintenance for the platform and a banner (at top) seemed more appropriate than sidebar which this feature currently had implemented.

Updated the CMS `/maintenance/announcements/` page to account for `Promote to Banner` behavior when needing to display at the top of the page.