fix: remove unique constraint on edx_user
With the auto linking of edx_user with the lti_users, the scenario where
multiple LTI consumers will create independent LtiUsers depending on the
same edx_user is created. This will not be possible if the edx_user has
a one-to-one relationship with the lti_user. This commit replaces the
one-to-one relationship with an one-to-many relationship so that
multiple LtiUser objects can be created referencing the same edx_user.