feat: Remove course outline question count and estimates by default.
We are temporarily disabling the course outline question count and estimate values by default until we can ensure that the values are correct.

The question count is off with Qualtrics XBlock indicating 1 Question when it has 10 Questions. We'll need to update the XBlock to return the number of questions on the Qualtrics survey.

The estimate metrics are off because we need to account for most course components. Also, we're improving the estimate values with our own override estimate calculations.