feat: finished multi-course export
assignment folders imsmanifest and module_meta now generates using multiple courses, changed identifier dictionary keys of sequentials and chapters to use a serialized object with custom hash function
Update for use of a new serialized object for chapters and sequentials due a weird issue where iterating through the same list of courselike keys in a different function led to slightly different chapters and sequentials causing for non-consistent keys, maybe due to memory of the objects?
Naming conventions were also added such as adding the tag at the end of the course key to the beginning of modules, etc.